Export hipfort CMake targets (#122)
If the ROCm math libraries are available at configuration-time, the
hipfort project will generate CMake targets for their fortran bindings.
To use the hipfort bindings, use

    find_package(hipfort COMPONENTS <libraries>)

where <libraries> is a space-delimited list of the C package names
(e.g., hip rocblas). The hipfort targets this provides are named the
same as the targets from the C libraries, but are found in the hipfort
namespace (e.g., hip::host becomes hipfort::host and roc::rocblas
becomes hipfort::rocblas).

At the moment, targets are only provided for the AMD backend.