Add CTest support (#130)
To build the tests with CMake add -DBUILD_TESTING=ON. After the build
is finished, the tests can be run with make test. e.g.

    cmake -S. -Bbuild -DBUILD_TESTING=ON
    make -C build
    CTEST_OUTPUT_ON_FAILURE=1 make -C build test

The tests can't successfully build with -std=f2003 or -std=f2008
because they depend on GNU extensions. I may look into making them more
portable in the future, but for the moment, enabling BUILD_TESTING
requires a compiler that supports -std=gnu.