Match subroutine names to C macro names
The ABI names are roctxMarkA, roctxRangePushA and roctxRangeStartA, but
in C these are exposed to users via macros that drop the A suffix. To
match the C interface as perceived by users, we can match the Fortran
subroutine name to the macro name rather than the actual symbol name.
Thank you to Brian Cornille and Dominic Charrier for this suggestion.